🧹 zoom: address review feedback
- role/group members: resolve through a per-connection user index instead
  of a per-member GetUser call (fixes N+1 on zoom.role.members and
  zoom.group.members)
- zoom.group settings: make the settings* fields lazy computed accessors
  backed by a double-check-locked fetch, mirroring zoom.account, so listing
  groups no longer issues a settings GET per group
- user.go: extract verified/ssoLinked derived predicates and add a
  table-driven decode_test.go
- client.go: User.Verified int64 -> int

+++ providers/zoom/connection/client.go
@@ -82,14 +82,15 @@ func (c *Client) get(ctx context.Context, path string, query url.Values, out any
 
 // User is the subset of Zoom's user object this provider reads.
 type User struct {
-	ID            string     `json:"id"`
-	Email         string     `json:"email"`
-	FirstName     string     `json:"first_name"`
-	LastName      string     `json:"last_name"`
-	DisplayName   string     `json:"display_name"`
-	Type          int64      `json:"type"`
-	Status        string     `json:"status"`
-	Verified      int64      `json:"verified"`
+	ID          string `json:"id"`
+	Email       string `json:"email"`
+	FirstName   string `json:"first_name"`
+	LastName    string `json:"last_name"`
+	DisplayName string `json:"display_name"`
+	Type        int64  `json:"type"`
+	Status      string `json:"status"`
+	// Verified is 0 or 1 (whether the user's email address is verified).
+	Verified      int        `json:"verified"`
 	LoginType     int64      `json:"login_type"`
 	RoleID        string     `json:"role_id"`
 	GroupIDs      []string   `json:"group_ids"`
